src/test/javascript/spec/app/entities/similarity/similarity-update.component.spec.ts
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { of } from 'rxjs';

import { ThinkItTestModule } from '../../../test.module';
import { SimilarityUpdateComponent } from 'app/entities/similarity/similarity-update.component';
import { SimilarityService } from 'app/entities/similarity/similarity.service';
import { Similarity } from 'app/shared/model/similarity.model';

describe('Component Tests', () => {
  describe('Similarity Management Update Component', () => {
    let comp: SimilarityUpdateComponent;
    let fixture: ComponentFixture<SimilarityUpdateComponent>;
    let service: SimilarityService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [ThinkItTestModule],
        declarations: [SimilarityUpdateComponent],
        providers: [FormBuilder],
      })
        .overrideTemplate(SimilarityUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(SimilarityUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(SimilarityService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new Similarity(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new Similarity();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
